+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All changes to ReportedIP Hive are documented here.
 
-## [2.1.41] — Unreleased
+## [2.1.41] — 2026-08-14
 
 ### New